dlx_rps/requirements.txt
pyqt5
pyqtgraph
zmq
matplotlib
numpy
pyyaml
scipy
pandas
pyarrow
pyyaml
coloredlogs